Prevent being able to call non-static methods on class objects
Previously, tests would fail if performed in a particular order, where a
Stack() was instantiated and then later Stack.push(...) was expected
to fail. An explicit test for this has been added.
This fix creates a new JavaMethod for every instantiated object, which
is probably not memory-optimal, but is a simple fix.